feat: add ProviderEnum and restructure LMStudio models with updated namespaces
