+++ Src/AiCommitMessage/Options/ProviderEnum.cs
@@ -1,0 +1,6 @@
+namespace AiCommitMessage.Options;
+
+public class ProviderEnum
+{
+    
+}
