+++ Tests/AiCommitMessage.Tests/Units/ProgramTests.cs
@@ -1,0 +1,6 @@
+namespace AiCommitMessage.Tests.Units;
+
+public class ProgramTests
+{
+    
+}
